rename post index to post/post, remove show post from nav. pr fixes!
